streamlit_frontend/requirements.txt
streamlit>=1.31
python-dotenv>=1.0
pydantic>=2.6
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40

langchain>=0.2.14
langchain-core>=0.2.34
langchain-openai>=0.1.20
langchain-community>=0.2.12
langgraph>=0.2.0

requests>=2.31
